Removed validations, Renamed files, Optimized operations on uploaded files
