+++ lib/components/Attachments/constants.js
@@ -22,7 +22,7 @@ export const DEFAULT_UPPY_CONFIG = {
 export const UPPY_UPLOAD_CONFIG = { formData: true, fieldName: "blob" };
 
 export const ATTACHMENT_OPTIONS = {
-  VIEW: "View",
+  DOWNLOAD: "Download",
   RENAME: "Rename",
   DELETE: "Delete",
 };
