--- lib/components/Attachments/utils.js
@@ -1,4 +1,0 @@
-export const validFileName = name =>
-  /^[a-zA-Z0-9._-\s]+.(jpg|jpeg|png|gif|mp4|pdf|xls|xlsx|csv)$/.test(name)
-    ? ""
-    : "File type is not permitted";
