+++ lib/index.scss
@@ -9,7 +9,7 @@
 @import "./styles/editor/media-uploader";
 @import "./styles/editor/emoji";
 @import "./styles/editor/media-editor";
-@import "./styles/components/file-attachment";
+@import "./styles/components/attachments";
 
 .ProseMirror {
   overflow-y: auto;
