--- lib/styles/components/_file-attachment.scss
@@ -1,77 +1,0 @@
-.ne-file-attachments-inner-wrapper {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-  margin-top: 0.5rem;
-  margin-bottom: 0.2rem;
-
-  .ne-file-attachment-wrapper {
-    display: flex;
-    padding-top: 0.25rem;
-    padding-bottom: 0.25rem;
-    font-size: 0.875rem;
-    line-height: 1.25rem;
-    justify-content: space-between;
-    align-items: center;
-    border-radius: 0.25rem;
-    border-width: 2px;
-    max-width: 20rem;
-  }
-
-  .ne-file-attachment-progress-wrapper {
-    display: flex;
-    gap: 1rem;
-    padding-top: 0.25rem;
-    padding-bottom: 0.25rem;
-    min-width: 20rem;
-  }
-
-  .ne-file-attachment-inner-wrapper {
-    display: flex;
-    align-items: center;
-    column-gap: 0.75rem;
-
-    .icon-opacity-75 {
-      opacity: 0.75;
-    }
-
-    .input-width {
-      width: 16rem;
-    }
-  }
-.drop-down-container {
-  margin-left: 1rem;
-
-  .menu-label {
-    text-transform: capitalize;
-  }
-}
-}
-
-.upload-btn {
-  margin-top: 0.5rem;
-}
-
-.truncate-ellipsis {
-  max-width: 15rem;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  display: block;
-}
-
-.truncate-ellipsis-progress {
-  max-width: 12rem;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  display: block;
-}
-
-.icon-opacity-50 {
-  opacity: 0.50;
-}
-
-input[type=file] {
-  display: none;
-}
